+++ src/main/kotlin/blackjack/model/card/CardDeck.kt
@@ -1,4 +1,4 @@
-package blackjack.model
+package blackjack.model.card
 
 class CardDeck(
     private val cards: List<Card>,
