+++ src/main/kotlin/blackjack/model/card/CardsStatus.kt
@@ -1,4 +1,4 @@
-package blackjack.model
+package blackjack.model.card
 
 enum class CardsStatus {
     BLACKJACK,
