+++ src/main/kotlin/blackjack/model/card/Denomination.kt
@@ -1,4 +1,4 @@
-package blackjack.model
+package blackjack.model.card
 
 enum class Denomination(
     val number: Int,
