+++ src/main/kotlin/blackjack/model/participant/Participant.kt
@@ -1,4 +1,7 @@
-package blackjack.model
+package blackjack.model.participant
+
+import blackjack.model.card.Card
+import blackjack.model.card.Cards
 
 abstract class Participant(
     val name: String,
