+++ src/main/kotlin/blackjack/model/participant/Players.kt
@@ -1,4 +1,6 @@
-package blackjack.model
+package blackjack.model.participant
+
+import blackjack.model.card.CardsStatus
 
 class Players(
     val value: List<Player>,
