+++ src/test/kotlin/blackjack/model/DealerTest.kt
@@ -1,5 +1,10 @@
 package blackjack.model
 
+import blackjack.model.card.Card
+import blackjack.model.card.CardShape
+import blackjack.model.card.Cards
+import blackjack.model.card.Denomination
+import blackjack.model.participant.Dealer
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertAll
