+++ src/test/kotlin/blackjack/model/DenominationTest.kt
@@ -1,5 +1,6 @@
 package blackjack.model
 
+import blackjack.model.card.Denomination
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
